pyproject.toml
[project]
name = "atlassian-backup-tool"
version = "0.1.0"
description = "atlassian cloud backup tool"
authors = [
    {name = "VinylStage",email = "mcpe9869@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "atlassian-python-api (>=4.0.7,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
